Add stopParsing flag to arg
When an arg is parsed that has the stopParsing flag set, the parser
will stop trying to parse flags, and treat the rest of the input as
arguments. One use of this feature is to support commands similar
to `yarn run` and `docker run`, as in `yarn run [<runFlags>] <script>
[<args>]`, where `<args>` are not documented or known to yarn, they are
simply passed on to the script being run.

The stopParsing flag allows an oclif command to behave similarly. You can
configure a positional argument to be the end of the input that is parsed
for flags, preventing oclif from detecting flags that occur afterward,
allowing the command to exclusively handle the rest of the input
through argv.